pyproject.toml
[project]
name = "fa-insight-harvester"
version = "0.1.0"
description = "失效分析周報結構化提取系統"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
    "python-multipart>=0.0.20",
    "python-pptx>=1.0.2",
    "Pillow>=11.1.0",
    "openai>=1.59.9",
    "jinja2>=3.1.5",
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "pgvector>=0.3.6",
    "PyJWT[crypto]>=2.10.1",
    "sse-starlette>=2.2.1",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28.1",
    "ruff>=0.8.0",
]
